extract execute_block, move simulate to Reactor, remove dead block_handler code
